dev: Persist caches used by PyOxidizer across builds
Reduces repeated downloads and significantly speeds up repeated builds.

I'd noticed the slowness when occasionally doing the local build or two,
but it had been a while since I was heavily rebuilding locally.  Doing
so again today motivated this restoration of the caching, which was
broken when we started building in a container for "Support even older
glibc (and thus Linux distro) versions for our standalone binary"
(b84d16c).