aws.s3: Refactor split_url() out of remote.s3 for use elsewhere
I want to use split_url() in runner.aws_batch.s3.bucket() for its
key/path handling.  The better error messaages are icing on the cake.

Adds an explicit assertion on url.scheme == "s3" to the top of the
function body since that was implicitly asserted by the previous sole
callers of the function.