runner.aws_batch: Support an optional prefix for S3 object keys
Provided after the bucket name in --aws-batch-s3-bucket or the
equivalent config field or env var.  For extra clarity when providing a
path, the bucket may now be specified as a URL (s3://), though, for
backwards compatibility, doing so is not required.

A key prefix allows for segmenting of the bucket objects and aids an
organization's organization.  This was requested by WA DOH in
discussions about their workflows.¹

An initially considered, but then discarded, alternative approach was to
add a new option separate from the bucket name, e.g.
--aws-batch-s3-prefix.  This ultimately felt unnecessary given I could
implement it within the existing option in a backwards-compatible way,
and, more importantly, like it'd end up in situations where a
wrong/unintended (bucket, prefix) combination was used because only one
option was changed instead of both together.

The targeted retry config introduced a while back² is removed and not
replaced because

 1. I don't think it's necessary?  boto3 has defaulted to three retries³
    for a long time, well before we added them explicitly, so I'm not
    even sure what effect our adding max_attempts had.

 2. It went against the approach of doing minimal AWS config ourselves
    and deferring to the config files and env vars supported by AWS
    libraries/tools themselves.  In this case, max_attempts in a config
    file or AWS_MAX_ATTEMPTS in the environment.⁴

¹ <https://bedfordlab.slack.com/archives/C01G4B4UY31/p1758159458179149>
  and subsequent notes in the Google Doc linked there.

² <https://github.com/nextstrain/cli/pull/80>
³ <https://boto3.amazonaws.com/v1/documentation/api/latest/guide/retries.html>
⁴ <https://boto3.amazonaws.com/v1/documentation/api/latest/guide/configuration.html>