Set deployment target and SDK
PyOxidizer targets¹ 10.9 for x86 (Intel), but that version is not
supported by any of the SDKs shipped with macOS 15.
MACOSX_DEPLOYMENT_TARGET is meant to override² this. It does not work,
so PyOxidizer tries and fails to find an SDK that supports 10.9. We get
around this by setting an SDK explicitly³ using an Xcode command⁴.

Even though MACOSX_DEPLOYMENT_TARGET seems to have no effect, I've kept
it to signal intent.

¹ https://pyoxidizer.readthedocs.io/en/stable/pyoxidizer_distributing_macos.html
² https://pyoxidizer.readthedocs.io/en/stable/pyoxidizer_distributing_macos.html#managing-portability-of-built-applications
³ https://pyoxidizer.readthedocs.io/en/stable/pyoxidizer_distributing_macos.html#build-machine-requirements
⁴ https://cmake.org/cmake/help/v4.2/variable/CMAKE_OSX_SYSROOT.html